requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
